[refactor](slqe/RuleTemplateForm): refactoring form hooks
