+++ packages/sqle/src/page/RuleTemplate/UpdateRuleTemplate/index.tsx
@@ -1,9 +1,7 @@
 import { useTranslation } from 'react-i18next';
-import { useBoolean, useRequest } from 'ahooks';
-import { useState, useCallback, useMemo, useEffect } from 'react';
-import { useNavigate, useParams } from 'react-router-dom';
+import { useCallback, useEffect } from 'react';
+import { useParams } from 'react-router-dom';
 import classNames from 'classnames';
-
 import { Space } from 'antd';
 import { PageLayoutHasFixedHeaderStyleWrapper } from '@actiontech/shared/lib/styleWrapper/element';
 import { BasicButton, PageHeader, BasicResult } from '@actiontech/shared';
@@ -12,80 +10,44 @@ import {
   IconSuccessResult
 } from '@actiontech/shared/lib/Icon/common';
 import RuleTemplateForm from '../RuleTemplateForm';
-
 import { useCurrentProject } from '@actiontech/shared/lib/global';
 import { RuleTemplateContStyleWrapper } from '../CreateRuleTemplate/style';
-import { useForm } from 'antd/es/form/Form';
-import { RuleTemplateBaseInfoFields } from '../RuleTemplateForm/BaseInfoForm/index.type';
-
 import rule_template from '@actiontech/shared/lib/api/sqle/service/rule_template';
 import { ResponseCode } from '@actiontech/shared/lib/enum';
+import { IRuleReqV1 } from '@actiontech/shared/lib/api/sqle/service/common';
 import {
-  IRuleProjectTemplateDetailResV1,
-  IRuleReqV1,
-  IRuleResV1
-} from '@actiontech/shared/lib/api/sqle/service/common';
-import EventEmitter from '../../../utils/EventEmitter';
-import EmitterKey from '../../../data/EmitterKey';
+  useUpdateRuleTemplateForm,
+  useBackToListPage
+} from '../../../hooks/useRuleTemplateForm';
 
 const UpdateRuleTemplate = () => {
   const { t } = useTranslation();
-  const navigate = useNavigate();
 
-  const [step, setStep] = useState(0);
   const { projectName, projectID } = useCurrentProject();
   const urlParams = useParams<{ templateName: string }>();
+
   const {
-    data: allRules,
-    loading: getAllRulesLoading,
-    run
-  } = useRequest((keyword?: string) =>
-    rule_template
-      .getRuleListV1({ fuzzy_keyword_rule: keyword })
-      .then((res) => res.data.data ?? [])
-  );
-
-  const [form] = useForm<RuleTemplateBaseInfoFields>();
-  const [dbType, setDbType] = useState<string>('');
-  const [ruleTemplate, setRuleTemplate] = useState<
-    IRuleProjectTemplateDetailResV1 | undefined
-  >();
-  const [activeRule, setActiveRule] = useState<IRuleResV1[]>([]);
-  const [databaseRule, setDatabaseRule] = useState<IRuleResV1[]>([]);
-  const [
+    step,
+    nextStep,
+    prevStep,
+    baseInfoFormSubmit,
+    form,
+    resetAll,
+    getAllRulesLoading,
+    ruleTemplate,
+    setRuleTemplate,
+    dbType,
+    activeRule,
+    setActiveRule,
+    databaseRule,
     updateTemplateLoading,
-    { setFalse: finishSubmit, setTrue: startSubmit }
-  ] = useBoolean(false);
-  const [baseInfoFormSubmitLoading, setBaseInfoFormSubmitLoading] =
-    useState(false);
-  const disabledForm = useMemo(
-    () => updateTemplateLoading,
-    [updateTemplateLoading]
-  );
-  const submitSuccessStatus = useMemo(() => step === 2, [step]);
+    finishSubmit,
+    startSubmit,
+    baseInfoFormSubmitLoading,
+    submitSuccessStatus
+  } = useUpdateRuleTemplateForm();
 
-  const baseInfoFormSubmit = useCallback(async () => {
-    setBaseInfoFormSubmitLoading(true);
-    try {
-      const values = await form.validateFields();
-      setDbType(values.db_type);
-      setDatabaseRule(
-        allRules?.filter((e) => e.db_type === values.db_type) ?? []
-      );
-      setStep(step + 1);
-      setBaseInfoFormSubmitLoading(false);
-    } catch (error) {
-      setBaseInfoFormSubmitLoading(false);
-    }
-  }, [form, step, allRules]);
-
-  const onGoRuleList = () => {
-    navigate(`/sqle/project/${projectID}/rule/template`);
-  };
-
-  const prevStep = useCallback(() => {
-    setStep(step - 1);
-  }, [step]);
+  const { onGotoRuleTemplateList } = useBackToListPage(projectID);
 
   const getRuleTemplate = useCallback(
     (fuzzyKeyword?: string) => {
@@ -103,25 +65,12 @@ const UpdateRuleTemplate = () => {
           }
         });
     },
-    [projectName, urlParams.templateName]
+    [projectName, urlParams.templateName, setActiveRule, setRuleTemplate]
   );
 
   useEffect(() => {
     getRuleTemplate();
   }, [getRuleTemplate]);
-
-  const resetAll = () => {
-    if (!ruleTemplate) return;
-    form.setFieldsValue({
-      templateName: ruleTemplate.rule_template_name,
-      templateDesc: ruleTemplate.desc,
-      db_type: ruleTemplate.db_type
-    });
-    setStep(0);
-    setActiveRule([...(ruleTemplate?.rule_list ?? [])]);
-    setDbType('');
-    EventEmitter.emit(EmitterKey.Search_Rule_Template_Rule_Clear_Value);
-  };
 
   const submit = useCallback(() => {
     startSubmit();
@@ -145,33 +94,13 @@ const UpdateRuleTemplate = () => {
       })
       .then((res) => {
         if (res.data.code === ResponseCode.SUCCESS) {
-          setStep(step + 1);
+          nextStep();
         }
       })
       .finally(() => {
         finishSubmit();
       });
-  }, [activeRule, finishSubmit, form, projectName, startSubmit, step]);
-
-  useEffect(() => {
-    if (dbType) {
-      const tempAllRules = allRules?.filter((e) => e.db_type === dbType) ?? [];
-      setDatabaseRule(tempAllRules);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [allRules]);
-
-  useEffect(() => {
-    const { unsubscribe } = EventEmitter.subscribe(
-      EmitterKey.Search_Rule_Template_Rule_Select_List,
-      (value) => {
-        run(value);
-        getRuleTemplate(value);
-      }
-    );
-
-    return unsubscribe;
-  }, [run, getRuleTemplate]);
+  }, [activeRule, finishSubmit, form, projectName, startSubmit, nextStep]);
 
   return (
     <>
@@ -179,20 +108,29 @@ const UpdateRuleTemplate = () => {
         <PageHeader
           fixed={step !== 1}
           title={
-            <BasicButton onClick={onGoRuleList} icon={<IconLeftArrow />}>
+            <BasicButton
+              onClick={onGotoRuleTemplateList}
+              icon={<IconLeftArrow />}
+            >
               {t('ruleTemplate.backToList')}
             </BasicButton>
           }
           extra={
             <Space size={12}>
               {step === 0 && (
-                <BasicButton onClick={resetAll} disabled={disabledForm}>
+                <BasicButton
+                  onClick={resetAll}
+                  disabled={updateTemplateLoading}
+                >
                   {t('common.reset')}
                 </BasicButton>
               )}
               {step === 1 && (
                 <>
-                  <BasicButton onClick={prevStep} disabled={disabledForm}>
+                  <BasicButton
+                    onClick={prevStep}
+                    disabled={updateTemplateLoading}
+                  >
                     {t('common.prevStep')}
                   </BasicButton>
                   <BasicButton
@@ -241,7 +179,7 @@ const UpdateRuleTemplate = () => {
               <BasicButton
                 type="primary"
                 key="view-rule-template-list"
-                onClick={onGoRuleList}
+                onClick={onGotoRuleTemplateList}
               >
                 {t('ruleTemplate.backToList')}
               </BasicButton>
